Added authorization with JWT and encrypt password

+++ backend/controllers/assignmentController.js
@@ -1,11 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Assignment = require('../models/assignmentModel')
+const User = require('../models/userModel');
+
 // get the assignments
 // GET api/assignments
 // private
 const getAssignments = asyncHandler(async (req, res) => {
-    const assignments = await Assignment.find()
+    const assignments = await Assignment.find({user: req.user.id})
     res.status(200).json(assignments)
 })
 
@@ -20,7 +22,8 @@ const setAssignment = asyncHandler(async (req, res) => {
     }
 
     const assignment = await Assignment.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(assignment)
@@ -37,6 +40,20 @@ const updateAssignment = asyncHandler(async (req, res) => {
         throw new Error('Assignment not found');
     }
 
+    const user = await User.findById(req.user.id);
+
+    // check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // ensure the logged in user matches the assignment user
+    if(assignment.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized.')
+    }
+
     const updatedAssignment = await Assignment.findByIdAndUpdate(req.params.id, req.body, {new:true})
     res.status(200).json(updateAssignment);
 })
@@ -50,6 +67,20 @@ const deleteAssignment = asyncHandler(async (req, res) => {
     if (!assignment) {
         res.status(400)
         throw new Error('Assignment not found');
+    }
+
+    const user = await User.findById(req.user.id);
+
+    // check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // ensure the logged in user matches the assignment user
+    if(assignment.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized.')
     }
 
     await assignment.deleteOne();
